SERVER ERROR 505 when uploading. I revised the validation for attachments on the create controller so they match the same validation on the uploadFiles controller.
In the storeAs() method on both the create and upload files controller, I changed the last parameter, which indicates the driver to use, from S3-public to S3.

For the URL variable on both the create and upload files controller,, I change the environment variable to point to AWS_BUCKET rather than AWS_BUCKET_PUBLIC since we are not using the public driver.

I created a folder called 'attachments' within the AWS Bucket online.

I confirmed that the attachment is uploading to the AWS bucket to both the Livewire-temp folder and the attachments folder.  I also confirmed that the path file is saving correctly on the database table for attachments.

I cannot view the attachment because I don't have the correct permissions assisned on the bucket.  I opened up another issue for that.